- Updated gitignore - Updated code, was generating too much unnecessary data. Should work better now.

+++ UnityFBXExporter/UnityFBXExporter/Vector2.h
@@ -3,6 +3,6 @@
 struct Vector2
 {
 public:
-	double X;
-	double Y;
+	float X;
+	float Y;
 };
